+++ lib/rabl/builder.rb
@@ -2,6 +2,13 @@ module Rabl
   class Builder
     include Rabl::Partials
 
+    SETTING_TYPES = {
+      :extends => :file,
+      :node    => :name,
+      :child   => :data,
+      :glue    => :data
+    } unless const_defined? :SETTING_TYPES
+
     # Constructs a new rabl hash based on given object and options
     # options = { :format => "json", :root => true, :child_root => true,
     #   :attributes, :node, :child, :glue, :extends }
@@ -29,44 +36,62 @@ module Rabl
     # compile_hash(:root_name => "user")
     def compile_hash(options={})
       @_result = {}
-      # Extends
-      @options[:extends].each do |settings|
-        extends(settings[:file], settings[:options], &settings[:block])
-      end if @options.has_key?(:extends)
-      # Attributes
-      @options[:attributes].each_pair do |attribute, settings|
-        attribute(attribute, settings)
-      end if @options.has_key?(:attributes)
-      # Node
-      @options[:node].each do |settings|
-        node(settings[:name], settings[:options], &settings[:block])
-      end if @options.has_key?(:node)
-      # Children
-      @options[:child].each do |settings|
-        child(settings[:data], settings[:options], &settings[:block])
-      end if @options.has_key?(:child)
-      # Glues
-      @options[:glue].each do |settings|
-        glue(settings[:data], settings[:options], &settings[:block])
-      end if @options.has_key?(:glue)
+      update_settings(:extends)
+      update_attributes
+      update_settings(:node)
+      update_settings(:child)
+      update_settings(:glue)
 
-      # Wrap result in root
-      if options[:root_name].present?
-        @_root_name = options[:root_name]
-      else # no root
-        @_root_name = nil
-      end
+      wrap_result(options[:root_name])
 
-      # Replace nil values with empty strings if configured
-      if Rabl.configuration.replace_nil_values_with_empty_strings
-        @_result = @_result.inject({}) do |hash, (k, v)|
-          hash[k] = v.nil? ? '' : v
-          hash
-        end
-      end
+      replace_nil_values          if Rabl.configuration.replace_nil_values_with_empty_strings
+      replace_empty_string_values if Rabl.configuration.replace_empty_string_values_with_nil_values
+      remove_nil_values           if Rabl.configuration.exclude_nil_values
 
       # Return Results
       @_root_name ? { @_root_name => @_result } : @_result
+    end
+
+    def replace_nil_values
+      @_result = @_result.inject({}) do |hash, (k, v)|
+        hash[k] = v.nil? ? '' : v
+        hash
+      end
+    end
+
+    def replace_empty_string_values
+      @_result = @_result.inject({}) do |hash, (k, v)|
+        hash[k] = v.try(:empty?) ? nil : v
+        hash
+      end
+    end
+
+    def remove_nil_values
+      @_result = @_result.inject({}) do |hash, (k, v)|
+        hash[k] = v unless v.nil?
+        hash
+      end
+    end
+
+    def wrap_result(root_name)
+      if root_name.present?
+        @_root_name = root_name
+      else # no root
+        @_root_name = nil
+      end
+    end
+
+    def update_settings(type)
+      settings_type = SETTING_TYPES[type]
+      @options[type].each do |settings|
+        send(type, settings[settings_type], settings[:options], &settings[:block])
+      end if @options.has_key?(type)
+    end
+
+    def update_attributes
+      @options[:attributes].each_pair do |attribute, settings|
+        attribute(attribute, settings)
+      end if @options.has_key?(:attributes)
     end
 
     # Indicates an attribute or method should be included in the json output
@@ -99,9 +124,11 @@ module Rabl
     # child(@users => :people) { ... }
     def child(data, options={}, &block)
       return false unless data.present? && resolve_condition(options)
-      name, object = data_name(data), data_object(data)
+      name   = is_name_value?(options[:root]) ? options[:root] : data_name(data)
+      object = data_object(data)
       include_root = is_collection?(object) && options.fetch(:object_root, @options[:child_root]) # child @users
       engine_options = @options.slice(:child_root).merge(:root => include_root)
+      engine_options.merge!(:object_root_name => options[:object_root]) if is_name_value?(options[:object_root])
       object = { object => name } if data.respond_to?(:each_pair) && object # child :users => :people
       @_result[name] = self.object_to_hash(object, engine_options, &block)
     end
@@ -118,9 +145,24 @@ module Rabl
     # Extends an existing rabl template with additional attributes in the block
     # extends("users/show") { attribute :full_name }
     def extends(file, options={}, &block)
+      return unless resolve_condition(options)
       options = @options.slice(:child_root).merge(:object => @_object).merge(options)
       result = self.partial(file, options, &block)
       @_result.merge!(result) if result.is_a?(Hash)
+    end
+
+    # Evaluate conditions given a symbol to evaluate
+    def call_condition_proc(condition, object, &blk)
+      blk = lambda { |v| v } unless block_given?
+      if condition.respond_to?(:call)
+        # condition is a block to pass to the block
+        blk.call(condition.call(object))
+      elsif condition.is_a?(Symbol) && object.respond_to?(condition)
+        # condition is a property of the object
+        blk.call(object.send(condition))
+      else
+        false
+      end
     end
 
     # resolve_condition(:if => true) => true
@@ -128,8 +170,14 @@ module Rabl
     # resolve_condition(:unless => lambda { |m| true }) => true
     def resolve_condition(options)
       return true if options[:if].nil? && options[:unless].nil?
-      result = options[:if] == true || (options[:if].respond_to?(:call) && options[:if].call(@_object)) if options.has_key?(:if)
-      result = options[:unless] == false || (options[:unless].respond_to?(:call) && !options[:unless].call(@_object)) if options.has_key?(:unless)
+      result = nil
+      if options.has_key?(:if)
+        result = options[:if] == true || call_condition_proc(options[:if], @_object)
+      end
+      if options.has_key?(:unless)
+        inverse_proc = lambda { |r| !r }
+        result = options[:unless] == false || call_condition_proc(options[:unless], @_object, &inverse_proc)
+      end
       result
     end