+++ test/configuration_test.rb
@@ -14,6 +14,8 @@ context 'Rabl::Configuration' do
     asserts(:json_engine).equals { json_engine }
     asserts(:cache_engine).is_a?(Rabl::CacheEngine)
     asserts(:replace_nil_values_with_empty_strings).equals false
+    asserts(:replace_empty_string_values_with_nil_values).equals false
+    asserts(:exclude_nil_values).equals false
   end
 
   context 'custom JSON engine configured as Symbol' do
@@ -55,4 +57,24 @@ context 'Rabl::Configuration' do
 
     asserts(:replace_nil_values_with_empty_strings).equals true
   end # replace nil values with empty strings
+
+  context 'replace empty string values with nil values' do
+    setup do
+      Rabl.configure do |c|
+        c.replace_empty_string_values_with_nil_values = true
+      end
+    end
+
+    asserts(:replace_empty_string_values_with_nil_values).equals true
+  end # replace empty string values with nil values
+
+  context 'exclude nil values' do
+    setup do
+      Rabl.configure do |c|
+        c.exclude_nil_values = true
+      end
+    end
+
+    asserts(:exclude_nil_values).equals true
+  end # exclude nil values
 end